requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-restframework==0.0.1
djangorestframework==3.16.0
docopt==0.6.2
et_xmlfile==2.0.0
idna==3.10
Js2Py==0.74
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pipwin==0.5.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.1.0
python-nmap==0.7.1
requests==2.32.4
scapy==2.6.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
